アセットインポーター令和最新版作成

+++ Assets/Scripts/Editor/AssetImporter/EditorButton.cs
@@ -2,31 +2,28 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
-using System.Threading.Tasks;
 using Cysharp.Threading.Tasks;
-using UniRx;
 using UnityEngine;
-
-#if UNITY_EDITOR
 using UnityEditor;
-#endif
 
 public class EditorButton : EditorWindow
 {
     private readonly CancellationTokenSource _cts = new();
     private CancellationToken _ct;
 
-    private ReactiveProperty<bool> _syncReleases = new();
-    private string _importText;
-    
+    private string _syncWaitingText = "リソース読み込み中";
+    private string _syncImportingText = "インポート実行待機中";
+
     private int _releasesSelectedIndex;
     private List<string> _releasesList = new();
-    
+
     private bool _showImportWindow;
     private bool _isInitialized = true;
     private Color _defaultLabelColor;
 
     private AssetsImporter _importer;
+
+    private bool _isFileChecking;
     // まめちしき
     // Unityには UnityEditor.AssetImporter というテクスチャ等のアセットを自動でインポートするやつがあるらしいわよ
     // https://light11.hatenadiary.com/entry/2018/04/05/194303
@@ -54,7 +51,7 @@ public class EditorButton : EditorWindow
             DrawColorLabel("初期化中...", Color.red);
             return;
         }
-        
+
         DrawColorLabel("インポートが終わるまでUnityのシーンを再生しないでください！", Color.red);
 
         // UnityPackageをインポートするときのファイル選択ウィンドウを表示するかのフラグ (デフォルトは表示しない false)
@@ -64,42 +61,42 @@ public class EditorButton : EditorWindow
         EditorGUILayout.LabelField(_showImportWindow
             ? "現在、アセットインポート時にインポートするフォルダを選べます"
             : "現在、自動的に全てのアセットの中身がインポートされます");
-        
+
         _releasesSelectedIndex = EditorGUILayout.Popup("パッケージバージョン", _releasesSelectedIndex, _releasesList.ToArray());
-        
+
         GUILayout.Space(10);
         GUILayout.Label("Selected: " + _releasesList[_releasesSelectedIndex]);
-        
-        if (GUILayout.Button("インポート"))
-        {
-            _ = DownLoad();
-        }
 
         GUILayout.Space(10);
 
-        if (GUILayout.Button("Sync Releases"))
+        if (GUILayout.Button("アセットのインポート"))
         {
-            _syncReleases.Value = false;
-            _ = Sync();
+            _syncImportingText = "アセットインポート中";
+            _ = Import(result => _syncImportingText = result);
         }
 
-        GUILayout.Label(_importText);
+        GUILayout.Label(_syncImportingText);
 
         GUILayout.Space(10);
 
-        if (GUILayout.Button("Assets"))
+        if (GUILayout.Button("アセット一覧の同期"))
         {
-            _ = Import();
+            _syncWaitingText = "リソース読み込み中";
+            _ = Sync(result => _syncWaitingText = result);
         }
+
+        GUILayout.Label(_syncWaitingText);
+
+        GUILayout.Space(50);
+        GUILayout.Label("-----------------これより下はプログラマー用-----------------");
+        _isFileChecking = GUILayout.Toggle(_isFileChecking, "ファイルの存在確認");
     }
 
     private void OnEnable()
     {
         _ct = _cts.Token;
 
-        _syncReleases.Subscribe(completed => _importText = completed ? "リソース読み込みが完了しました" : "リソース読み込み待機中");
-        
-        _ = Sync();
+        _ = Sync(result => _syncWaitingText = result);
     }
 
     private void OnDisable()
@@ -108,40 +105,54 @@ public class EditorButton : EditorWindow
         _importer.Dispose();
     }
 
-    private async UniTaskVoid Sync()
+    private async UniTaskVoid Sync(Action<string> callback)
     {
-        _importer = new AssetsImporter();
-        await _importer.GetReleases("releases", _ct);
-        _releasesList.Clear();
-        foreach (var release in _importer.Releases)
-        {
-            _releasesList.Add(release.TagName);
-        }
-        _syncReleases.Value = true;
-    }
-
-    private async UniTaskVoid Import()
-    {
-        var id = _importer.Releases[_releasesSelectedIndex].Assets[0].ID;
-        
-        await _importer.GetAssetUrl("asset", id, _ct);
-    }
-
-    private async Task DownLoad()
-    {
-        // UnityPackageの存在チェックはAssetsImporter.csで行うので、ここではやらない
-
-        var task = _importer.StartFetching();
-        var result = await task;
-
-        if (string.IsNullOrEmpty(result))
-        {
-            Debug.LogError("ファイルのパスが取得できませんでした");
-            return;
-        }
-
         try
         {
+            _importer = new AssetsImporter();
+            await _importer.GetReleases("releases", _ct);
+            _releasesList.Clear();
+            foreach (var release in _importer.Releases)
+            {
+                _releasesList.Add(release.TagName);
+            }
+        }
+        catch (Exception e)
+        {
+            Debug.LogError($"Releases Syncing Exception: {e}");
+        }
+
+        callback("リソース読み込みが完了しました");
+    }
+
+    private async UniTaskVoid Import(Action<string> callback)
+    {
+        try
+        {
+            var id = _importer.Releases[_releasesSelectedIndex].Assets[0].ID;
+
+            var filePath = await _importer.GetAsset("asset", id, _ct);
+            Download(filePath);
+            callback("アセットのインポートが完了しました");
+        }
+        catch (Exception e)
+        {
+            callback("アセットのインポートに失敗しました");
+            Debug.LogError($"Importing Exception: {e}");
+            throw;
+        }
+    }
+
+    private void Download(string result)
+    {
+        try
+        {
+            if (string.IsNullOrEmpty(result))
+            {
+                Debug.LogError("ファイルのパスが取得できませんでした");
+                return;
+            }
+
             var files = Directory.GetFiles(result, "*.unitypackage", SearchOption.AllDirectories);
 
             foreach (var file in files)
@@ -151,14 +162,22 @@ public class EditorButton : EditorWindow
         }
         catch (Exception e)
         {
-            Debug.LogError($"Exception {e}");
+            Debug.LogError($"Download Exception: {e}");
         }
-        finally
+
+        if (_isFileChecking)
         {
-            Directory.Delete(result, true);
+            return;
         }
+
+        Directory.Delete(result, true);
     }
 
+    /// <summary>
+    /// GUI.Labelの色を変更して表示するヘルパーメソッド
+    /// </summary>
+    /// <param name="text">色を変更したいテキスト</param>
+    /// <param name="color">変更したい色</param>
     private void DrawColorLabel(string text, Color color)
     {
         var style = GUI.skin.label;
